Address reviewer feedback: move XSAMultiheadAttention to examples, assert HAS_FLEX
- Remove XSAMultiheadAttention from attn_gym/mods (not prescriptive on nn.Module)
- Move class into examples/xsa_attention.py where it belongs
- Add assert HAS_FLEX in example_composable() instead of silent SDPA fallback
- Update test import to pull XSAMultiheadAttention from examples
