+++ attn_gym/mods/__init__.py
@@ -1,4 +1,4 @@
 from attn_gym.mods.activation import generate_activation_score_mod, undo_softmax
 from attn_gym.mods.alibi import generate_alibi_bias
 from attn_gym.mods.softcapping import generate_tanh_softcap
-from attn_gym.mods.exclusive_sa import exclusive_output_mod, XSAMultiheadAttention
+from attn_gym.mods.exclusive_sa import exclusive_output_mod
